+++ src/context.rs
@@ -6,11 +6,9 @@ use std::task::Poll;
 use std::{mem, task};
 
 use futures_core::future::BoxFuture;
-use futures_core::FusedFuture;
 use futures_util::FutureExt;
 
 use crate::envelope::{HandlerSpan, Shutdown};
-use crate::inbox::rx::{ReceiveFuture as InboxReceiveFuture, RxStrong};
 use crate::inbox::ActorMessage;
 use crate::{Actor, Mailbox};
 
@@ -49,41 +47,6 @@ impl<'m, A: Actor> Context<'m, A> {
     /// Get a reference to the [`Mailbox`] of this actor.
     pub fn mailbox(&mut self) -> &mut Mailbox<A> {
         self.mailbox
-    }
-}
-
-/// A message sent to a given actor, or a notification that it should shut down.
-pub struct Message<A>(pub(crate) ActorMessage<A>);
-
-/// A future which will resolve to the next message to be handled by the actor.
-#[must_use = "Futures do nothing unless polled"]
-pub struct ReceiveFuture<A>(pub(crate) InboxReceiveFuture<A, RxStrong>);
-
-impl<'c, A> ReceiveFuture<A> {
-    /// Cancel the receiving future, returning a message if it had been fulfilled with one, but had
-    /// not yet been polled after wakeup. Future calls to `Future::poll` will return `Poll::Pending`,
-    /// and `FusedFuture::is_terminated` will return `true`.
-    ///
-    /// This is important to do over `Drop`, as with `Drop` messages may be sent back into the
-    /// channel and could be observed as received out of order, if multiple receives are concurrent
-    /// on one thread.
-    #[must_use = "If dropped, messages could be lost"]
-    pub fn cancel(&mut self) -> Option<Message<A>> {
-        self.0.cancel().map(Message)
-    }
-}
-
-impl<A> Future for ReceiveFuture<A> {
-    type Output = Message<A>;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Self::Output> {
-        self.0.poll_unpin(cx).map(Message)
-    }
-}
-
-impl<A> FusedFuture for ReceiveFuture<A> {
-    fn is_terminated(&self) -> bool {
-        self.0.is_terminated()
     }
 }
